Send a packet to make the depot available on client.
